backend/approval-processing-service/requirements.txt
fastapi
uvicorn[standard]

grpcio
protobuf
httpx
aio-pika>=9.3.0,<10
